Updated index.js and removed duplicate server file server.js
